refactor: generic meetup component
