+++ frontend/src/pages/mod.rs
@@ -1,1 +1,2 @@
+pub mod bars;
 pub mod tech_events;
